+++ qiskit/providers/ibmq/api_v2/__init__.py
@@ -5,7 +5,6 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""REST adaptors for the IBM Q Api version 2."""
+"""IBM Q API version 2 connector and utilities."""
 
-from .auth import Auth
-from .root import Api
+from .ibmqclient import IBMQClient
