+++ qiskit/providers/ibmq/api_v2/rest/auth.py
@@ -7,7 +7,7 @@
 
 """Auth REST adaptor for the IBM Q Api version 2."""
 
-from qiskit.providers.ibmq.api.rest.base import RestAdaptorBase
+from .base import RestAdaptorBase
 
 
 class Auth(RestAdaptorBase):
