+++ qiskit/providers/ibmq/api_v2/rest/job.py
@@ -9,7 +9,7 @@
 
 import json
 
-from qiskit.providers.ibmq.api.rest.base import RestAdaptorBase
+from .base import RestAdaptorBase
 from .util import build_url_filter
