+++ qiskit/providers/ibmq/api_v2/rest/util.py
